encoding: Handle mail body with wrong Content-Type
A mail client may send an email with a header like

Content-Type: text/plain; charset=utf-8

But there is technically no enforcement police to actually check or
detect that the body is actually encoded as written. If the body is actually
encoded using ISO-8859-1, it will contain "invalid" UTF-8 characters, which
may cause htmlentities() to blank out the entire text body.

This patch allows htmlentities() to simply drop the "invalid" characters
while keeping the rest of the body. PHP documentation recommends against the
flag; however, we are generally relying on htmLawed to clean message
content.

The example email in our case was from an unknown mail client; however, it
also contained a X-MS-TNEF-Correlator header... The tell-tale sign was an
&nbsp; character encoded as "0xa0" rather than the utf-8 version "0xc2a0"