orm: Use local side of relationship if possible
This attempts to use the local side of a relationship when comparing a
relationship to a model instance. For instance:

    User::objects()->filter(['org' => $org]);

The previous code would join through the organization table, which could
possibly not work (e.g. if the org had been just deleted).

This patch will attempt to use the local side of the relationship (the side
on the User model) and avoid the join through the organization table.